+++ src/shared/api/adapters/axios/AxiosAdapter.ts
