+++ src/shared/api/adapters/fetch/FetchAdapter.ts
@@ -1,8 +1,8 @@
 import { API } from "../../../../app.constants";
-import { HttpClientResponse } from "../../types/api-adapter.types";
+import { HttpClientResponse } from "../../types/ApiAdapter.types";
 import FetchErrorHandler from "./utils/fetch-error-handler";
 
-class FetchAdapter {
+export default class FetchAdapter {
   async get<T>(endpoint: string): Promise<HttpClientResponse<T>> {
     const response = await fetch(`${API.BASE_URL}${endpoint}`, {
       method: "GET",
@@ -14,5 +14,3 @@ class FetchAdapter {
     return { data };
   }
 }
-
-export default FetchAdapter;
