+++ src/shared/api/types/ApiAdapter.types.ts
